Address coefficients by caller-visible factor labels
Introduce FactorEncoding as the translation boundary between caller labels and internal compact positions. Keep u32 labels in public result APIs and translate unidentified directions at the solver boundary.

+++ crates/within/src/solver/tests.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 /// in `tests/slopes_routing.rs`. A positive slope-only term is not centered by
 /// whitening, so the signed pair stays all-positive — balanced — while generic
 /// `z` keeps it strictly inside the PSD cone: genuine surplus, grounded.
-fn at(term: usize, level: usize, column: usize) -> CoefficientAddress {
+fn at(term: usize, level: u32, column: usize) -> CoefficientAddress {
     CoefficientAddress {
         channel: Channel { term, column },
         level,
